feat(ring_theory): the comap of an ideal is not zero
Let `f` be a ring homomorphism from `R` to `S` and `I` be an ideal in `S`.
To show that `I.comap f` is not the zero ideal, we can show `I` contains
a non-zero root of some non-zero polynomial `p : polynomial R`.
As a corollary, if `S` is algebraic over `R`, nonzero ideals in `S` lie over
nonzero ideals in `R`.